refactor: Выделить экшны редьюсеров в отдельные файлы
