+++ src/services/slices/index.js
@@ -1,8 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
-import ingredients from './ingredientsSlice';
-import burger from './burgerSlice';
-import order from './orderSlice';
-import auth from './authSlice';
+import ingredients from './ingredients';
+import burger from './burger';
+import order from './order';
+import auth from './auth';
 
 export default configureStore({
   reducer: {
